Add client mod
Loading Interdimensional Portals on the client will now automatically obtain portal positions from the server and tint them accordingly.

There is an issue with non-custom portals at the moment regarding their colour: there is a fixme and is as simple as replacing the value 1908001 with whatever the vanilla portal uses.

Signed-off-by: yitzy299 <yitzy299@users.noreply.github.com>